requirements.txt
feedparser
requests
python-dotenv
groq